Refactor core and spring packages
