+++ org-sync-core/src/main/java/org/orgsync/core/spec/YamlSpecLoader.java
@@ -1,4 +1,4 @@
-package org.orgsync.core;
+package org.orgsync.core.spec;
 
 import java.io.IOException;
 import java.io.InputStream;
