+++ org-sync-core/src/main/java/org/orgsync/core/state/SyncStateRepository.java
@@ -1,4 +1,4 @@
-package org.orgsync.core;
+package org.orgsync.core.state;
 
 import java.util.Optional;
