+++ org-sync-spring-sample/src/main/java/org/orgsync/springsample/OrgSyncSpringSample.java
@@ -1,16 +1,16 @@
 package org.orgsync.springsample;
 
-import org.orgsync.core.DomainEvent;
-import org.orgsync.core.DomainEventPublisher;
-import org.orgsync.core.LockManager;
-import org.orgsync.core.OrgChartClient;
-import org.orgsync.core.SyncEngine;
-import org.orgsync.core.SyncResponse;
-import org.orgsync.core.SyncStateRepository;
-import org.orgsync.core.YamlSyncSpec;
-import org.orgsync.spring.InMemoryLockManager;
-import org.orgsync.spring.OrgSyncConfiguration;
-import org.orgsync.spring.SpringDomainEventPublisher;
+import org.orgsync.core.client.OrgChartClient;
+import org.orgsync.core.engine.SyncEngine;
+import org.orgsync.core.engine.SyncResponse;
+import org.orgsync.core.event.DomainEvent;
+import org.orgsync.core.event.DomainEventPublisher;
+import org.orgsync.core.lock.LockManager;
+import org.orgsync.core.state.SyncStateRepository;
+import org.orgsync.core.spec.YamlSyncSpec;
+import org.orgsync.spring.config.OrgSyncConfiguration;
+import org.orgsync.spring.event.SpringDomainEventPublisher;
+import org.orgsync.spring.lock.InMemoryLockManager;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
